release 0.0.4: 我认为这个问题应该是由于上升和下降沿的diff造成的，我预计再下一个版本完成更加详细的上升和下降的计算。在本版本之中引入了dfs解决了时序路径打印的问题，同时将ista的liberty parser 引入到项目之中，可以成功的解析实际使用的liberty

+++ src/sta.cpp
@@ -998,8 +998,7 @@ double STAWorker::process_endpoint_timing(TimingEndpoint &ep,
               arc.fall_constraint.value(), data_trans, clk_trans,
               t->variable_1.value(), t->variable_2.value());
       }
-      ep.Setup_req = static_cast<int>(std::max(setup_rise, setup_fall) *
-                                      NS_TO_PS); // ns -> ps
+      ep.Setup_req = std::max(setup_rise, setup_fall) * NS_TO_PS; // ns -> ps
     }
     if (arc.timing_type == celllib::TimingType::HOLD_RISING ||
         arc.timing_type == celllib::TimingType::HOLD_FALLING) {
@@ -1022,8 +1021,7 @@ double STAWorker::process_endpoint_timing(TimingEndpoint &ep,
               arc.fall_constraint.value(), data_trans, clk_trans,
               t->variable_1.value(), t->variable_2.value());
       }
-      ep.Hold_req = static_cast<int>(std::max(hold_rise, hold_fall) *
-                                     NS_TO_PS); // ns -> ps
+      ep.Hold_req = std::max(hold_rise, hold_fall) * NS_TO_PS; // ns -> ps
     }
   }
   return ep.Setup_req.value_or(0);
